change: remove sdk.Router and sdk.Handler (and derivatives)
