+++ Makefile
@@ -315,7 +315,7 @@ test-cover:
 
 test-rosetta:
 	docker build -t rosetta-ci:latest -f contrib/rosetta/node/Dockerfile .
-	docker-compose -f contrib/rosetta/docker-compose.yaml up --abort-on-container-exit --exit-code-from test_rosetta --build
+	docker-compose -f contrib/rosetta/docker-compose.yaml up
 .PHONY: test-rosetta
 
 benchmark:
