--- x/evidence/keeper/querier.go
@@ -1,80 +1,0 @@
-package keeper
-
-import (
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/x/evidence/exported"
-	"github.com/cosmos/cosmos-sdk/x/evidence/types"
-
-	abci "github.com/tendermint/tendermint/abci/types"
-)
-
-func NewQuerier(k Keeper, legacyQuerierCdc *codec.LegacyAmino) sdk.Querier {
-	return func(ctx sdk.Context, path []string, req abci.RequestQuery) ([]byte, error) {
-		var (
-			res []byte
-			err error
-		)
-
-		switch path[0] {
-		case types.QueryEvidence:
-			res, err = queryEvidence(ctx, req, k, legacyQuerierCdc)
-
-		case types.QueryAllEvidence:
-			res, err = queryAllEvidence(ctx, req, k, legacyQuerierCdc)
-
-		default:
-			err = sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "unknown %s query endpoint: %s", types.ModuleName, path[0])
-		}
-
-		return res, err
-	}
-}
-
-func queryEvidence(ctx sdk.Context, req abci.RequestQuery, k Keeper, legacyQuerierCdc *codec.LegacyAmino) ([]byte, error) {
-	var params types.QueryEvidenceRequest
-
-	err := legacyQuerierCdc.UnmarshalJSON(req.Data, &params)
-	if err != nil {
-		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONUnmarshal, err.Error())
-	}
-
-	evidence, ok := k.GetEvidence(ctx, params.EvidenceHash)
-	if !ok {
-		return nil, sdkerrors.Wrap(types.ErrNoEvidenceExists, params.EvidenceHash.String())
-	}
-
-	res, err := codec.MarshalJSONIndent(legacyQuerierCdc, evidence)
-	if err != nil {
-		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
-	}
-
-	return res, nil
-}
-
-func queryAllEvidence(ctx sdk.Context, req abci.RequestQuery, k Keeper, legacyQuerierCdc *codec.LegacyAmino) ([]byte, error) {
-	var params types.QueryAllEvidenceParams
-
-	err := legacyQuerierCdc.UnmarshalJSON(req.Data, &params)
-	if err != nil {
-		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONUnmarshal, err.Error())
-	}
-
-	evidence := k.GetAllEvidence(ctx)
-
-	start, end := client.Paginate(len(evidence), params.Page, params.Limit, 100)
-	if start < 0 || end < 0 {
-		evidence = []exported.Evidence{}
-	} else {
-		evidence = evidence[start:end]
-	}
-
-	res, err := codec.MarshalJSONIndent(legacyQuerierCdc, evidence)
-	if err != nil {
-		return nil, sdkerrors.Wrap(sdkerrors.ErrJSONMarshal, err.Error())
-	}
-
-	return res, nil
-}
